Remove useless condition
