+++ www/front_src/src/Main/index.tsx
@@ -14,9 +14,10 @@ import weekday from 'dayjs/plugin/weekday';
 import isBetween from 'dayjs/plugin/isBetween';
 import isSameOrBefore from 'dayjs/plugin/isSameOrBefore';
 import duration from 'dayjs/plugin/duration';
-import { isNil } from 'ramda';
-import { Route, Routes } from 'react-router-dom';
+import { and, isNil, not } from 'ramda';
+import { Route, Routes, useNavigate } from 'react-router-dom';
 import { useAtomValue } from 'jotai/utils';
+import { useAtom } from 'jotai';
 
 import { platformInstallationStatusAtom } from '../platformInstallationStatusAtom';
 import reactRoutes from '../reactRoutes/routeMap';
@@ -24,6 +25,7 @@ import reactRoutes from '../reactRoutes/routeMap';
 import Provider from './Provider';
 import { MainLoaderWithoutTranslation } from './MainLoader';
 import useMain from './useMain';
+import { areUserParametersLoadedAtom } from './useUser';
 
 dayjs.extend(localizedFormat);
 dayjs.extend(utcPlugin);
@@ -41,11 +43,46 @@ const ResetPasswordPage = React.lazy(() => import('../ResetPassword'));
 const AppPage = React.lazy(() => import('./InitializationPage'));
 
 const Main = (): JSX.Element => {
+  const navigate = useNavigate();
+
   useMain();
 
+  const [areUserParametersLoaded] = useAtom(areUserParametersLoadedAtom);
   const platformInstallationStatus = useAtomValue(
     platformInstallationStatusAtom,
   );
+
+  const navigateTo = (path: string): void => {
+    navigate(path);
+    window.location.reload();
+  };
+
+  React.useEffect(() => {
+    if (isNil(platformInstallationStatus) || isNil(areUserParametersLoaded)) {
+      return;
+    }
+
+    if (not(platformInstallationStatus.isInstalled)) {
+      navigateTo(reactRoutes.install);
+
+      return;
+    }
+
+    const canUpgrade = and(
+      platformInstallationStatus.hasUpgradeAvailable,
+      not(areUserParametersLoaded),
+    );
+
+    if (canUpgrade) {
+      navigateTo(reactRoutes.upgrade);
+
+      return;
+    }
+
+    if (not(areUserParametersLoaded)) {
+      navigate(reactRoutes.login);
+    }
+  }, [platformInstallationStatus, areUserParametersLoaded]);
 
   if (isNil(platformInstallationStatus)) {
     return <MainLoaderWithoutTranslation />;
